pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fitz>=0.0.1.dev2",
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
]
